feat: Implement logic to log the build and run
Implement logic to log the build and run of every target and generate an overall build and run report.

+++ scripts/utils/tests_runner.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ -z "$1" ]; then
+  echo "Usage: ./tests_runner.sh <app_dir_path>"
+  exit 1
+fi
+
+app_dir_path="$1"
+
 rm -fr .tests/
 source testing-fw-venv/bin/activate
-python src/main.py src/tester_config.yaml
+python src/main.py "$app_dir_path"
